feat: 리프레쉬토큰에 ip주소까지 저장하도록 변경 (#178)

+++ src/main/java/com/flab/readnshare/domain/auth/service/AuthService.java
@@ -1,11 +1,13 @@
 package com.flab.readnshare.domain.auth.service;
 
+import com.flab.readnshare.domain.auth.domain.RefreshToken;
 import com.flab.readnshare.domain.auth.dto.SignInRequestDto;
 import com.flab.readnshare.domain.auth.repository.RefreshTokenRepository;
 import com.flab.readnshare.domain.member.domain.Member;
 import com.flab.readnshare.domain.member.repository.MemberRepository;
 import com.flab.readnshare.global.common.auth.jwt.JwtUtil;
 import com.flab.readnshare.global.common.exception.AuthException;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -49,21 +51,31 @@ public class AuthService {
     }
 
     // memberId에 대한 refresh 토큰을 생성 후 HTTP 응답 쿠키에 설정
-    public void sendRefreshToken(HttpServletResponse response, Long memberId) {
-        String refreshToken = jwtUtil.createRefreshToken(memberId);
+    public void sendRefreshToken(HttpServletRequest request, HttpServletResponse response, Long memberId) {
+        String clientIp = request.getRemoteAddr();
+        String refreshToken = jwtUtil.createRefreshToken(memberId, clientIp);
         jwtUtil.setRefreshTokenCookie(response, refreshToken);
     }
     // memberId에 대한 access, refresh 토큰을 생성 후 HTTP 응답 헤더와 쿠키에 설정
-    public void issueTokens(HttpServletResponse response, Long memberId) {
+    public void issueTokens(HttpServletRequest request, HttpServletResponse response, Long memberId) {
         sendAccessToken(response, memberId);
-        sendRefreshToken(response, memberId);
+        sendRefreshToken(request, response, memberId);
         log.info("AccessToken, RefreshToken 발급 완료 - memberId: {}", memberId);
     }
 
     // refresh 토큰을 삭제하고 새로운 access, refresh 토큰을 발급
-    public void updateRefreshToken(HttpServletResponse response, Long memberId, String oldRefreshToken) {
+    //토큰 재발급 시 클라이언트 IP가 저장된 IP와 일치하는지 검증
+    public void updateRefreshToken(HttpServletRequest request, HttpServletResponse response, Long memberId, String oldRefreshToken) {
+        RefreshToken storedToken = refreshTokenRepository.findById(oldRefreshToken)
+                .orElseThrow(AuthException.ExpiredTokenException::new);
+        String requestIp = request.getRemoteAddr();
+
+        if (!storedToken.getIpAddress().equals(requestIp)) {
+            throw new AuthException.DeniedTokenException();
+        }
+
         refreshTokenRepository.deleteById(oldRefreshToken);
-        issueTokens(response, memberId);
+        issueTokens(request, response, memberId);
     }
 
     // refresh 토큰을 검증
